Project/airflow/requirements.txt
apache-airflow-providers-google==8.11.0
#pyarrow==11.0.0
#dbt-bigquery==1.4.3
kaggle==1.5.13